src/api/picture-upload/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
Faker==13.3.4
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
